Add AlertDialogModal to the BoardColumns. Define new schema for short Job post.
